requirements.txt
fastapi
uvicorn
selenium
undetected-chromedriver
beautifulsoup4
tqdm
openai
pinecone-client
python-dotenv
